Fix compile with QT_NO_NARROWING_CONVERSIONS_IN_CONNECT flags
Summary: Fix compile with QT_NO_NARROWING_CONVERSIONS_IN_CONNECT flags

Reviewers: mgallien

Reviewed By: mgallien

Differential Revision: https://phabricator.kde.org/D15588

+++ src/audiowrapper.h
@@ -148,7 +148,7 @@ public Q_SLOTS:
 
     void stop();
 
-    void seek(int position);
+    void seek(qint64 position);
 
     void setAudioRole(QAudio::Role audioRole);
